# openapi.yaml
openapi: 3.0.0
info:
  title: Taska Backend API
  description: REST API для платформы Taska
  version: 1.0.0
  contact:
    name: Taska Support
    email: support@taska.com
servers:
  - url: http://localhost:3000
    description: Local server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        message:
          type: string
          description: Error message
        errors:
          type: array
          items:
            type: object
            properties:
              msg:
                type: string
              param:
                type: string
              location:
                type: string
...
    Column:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        order:
          type: integer
          minimum: 0
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: string
        updatedAt:
          type: string
          format: date-time
          nullable: true
        updatedBy:
          type: string
          nullable: true
...
paths:
...
  /api/projects/{projectId}/columns:
    get:
      summary: Получить список колонок проекта
      tags:
        - Columns
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: projectId
          schema:
            type: string
          description: ID проекта
      responses:
        '200':
          description: Список колонок
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Column'
        '401':
          description: Неавторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Нет доступа к проекту
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Проект не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      summary: Создать новую колонку
      tags:
        - Columns
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: projectId
          schema:
            type: string
          description: ID проекта
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                order:
                  type: integer
                  minimum: 0
      responses:
        '201':
          description: Колонка создана
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  columnId:
                    type: string
        '400':
          description: Ошибка валидации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Неавторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Только PM может создавать колонки
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Проект не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...